feat(leafy): improve shaders + 3D pyramid rotation + consolidated build
- Add continuous 3D rotation animation to vision pyramid (X/Y/Z axes)
- Enhance hero shader with gold flow lines, particles, ripples, glows
- Load all shaders from local .glsl files instead of embedded JS
- Implement lazy-loading shader gallery (click-to-play)
- Reduce WebGL contexts to avoid browser exhaustion
- Merge fix-imports.sh into build.sh (single build script)
- Fix "Technologie" typos to "Technology"

+++ b_Work_Profiles/leafy/src/shaders/starnest.glsl
@@ -1,11 +1,7 @@
-// Star Nest Shader - Adapted from Shadertoy by Kali
-// Original: https://www.shadertoy.com/view/XlfGRj
-
-precision highp float;
+precision mediump float;
 uniform vec2 resolution;
 uniform float time;
 uniform float intensity;
-
 #define iterations 17
 #define formuparam 0.53
 #define volsteps 20
@@ -17,44 +13,33 @@ uniform float intensity;
 #define darkmatter 0.300
 #define distfading 0.730
 #define saturation 0.850
-
 void main() {
     vec2 uv = gl_FragCoord.xy / resolution.xy - 0.5;
     uv.y *= resolution.y / resolution.x;
-
     vec3 dir = vec3(uv * zoom, 1.0);
     float t = time * speed + 0.25;
-
     vec3 from = vec3(1.0, 0.5, 0.5);
     from += vec3(t * 2.0, t, -2.0);
-
     // Volumetric rendering
     float s = 0.1, fade = 1.0;
     vec3 v = vec3(0.0);
-
     for(int r = 0; r < volsteps; r++) {
         vec3 p = from + s * dir * 0.5;
         p = abs(vec3(tile) - mod(p, vec3(tile * 2.0)));
-
         float pa, a = pa = 0.0;
         for(int i = 0; i < iterations; i++) {
             p = abs(p) / dot(p, p) - formuparam;
             a += abs(length(p) - pa);
             pa = length(p);
         }
-
         float dm = max(0.0, darkmatter - a * a * 0.001);
         a *= a * a;
-
         if(r > 6) fade *= 1.0 - dm;
-
         v += fade;
         v += vec3(s, s * s, s * s * s * s) * a * brightness * fade;
         fade *= distfading;
         s += stepsize;
     }
-
     v = mix(vec3(length(v)), v, saturation);
-
     gl_FragColor = vec4(v * 0.01 * intensity, 1.0);
 }
